data: started to integrate wtw, plus cleanup
