+++ apps/data-service/package.json
@@ -19,7 +19,9 @@
     "node-cron": "^3.0.0",
     "pm2": "^5.1.2",
     "@dills1220/tmdb": "2.7.1",
-    "dotenv": "~11.0.0"
+    "dotenv": "~11.0.0",
+    "@dills1220/wtw": "workspace:*",
+    "redis-sdk": "workspace:*"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
