--- apps/data-service/redis-client.ts
@@ -1,1 +1,0 @@
-//TODO implement Redis client here
